moved end_screen to new cpp
made header
fixed guard for project_window
