+++ fifteen_game.h
@@ -1,2 +1,3 @@
 #include "Tile_button.h"
 #include "Project_window.h"
+#include "End_screen.h"
