+++ usr/Scores_list_10.txt
@@ -6,23 +6,25 @@ YAY 80
 
 YAY 80
 
-YAY 20
+YAY 80
 
 YAY 20
 
 YAY 20
 
 YAY 20
+
+YAY 20
+
+-- 10
+
+-- 10
+
+-- 10
 
 YAY 10
 
 YAY 10
-
--- 10
-
--- 10
-
--- 10
 
 -- 10
